revert backspace deleting row on non-mac

+++ chrome/content/zotero/preferences/preferences_cite.jsx
@@ -125,7 +125,7 @@ Zotero_Preferences.Cite = {
 				}
 			];
 			var handleKeyDown = (event) => {
-				if (event.key == 'Delete' || event.key == 'Backspace') {
+				if (event.key == 'Delete' || Zotero.isMac && event.key == 'Backspace') {
 					Zotero_Preferences.Cite.deleteStyle();
 					return false;
 				}
